+++ docs/changelog.md
@@ -19,6 +19,10 @@
 
 ## 更新说明
 
+1.5.43 fix: fix stock_zt_pool_em interface
+
+    1. 修复 stock_zt_pool_em 接口，获取所有股票的数据
+
 1.5.42 fix: fix stock_zt_pool_dtgc_em interface
 
     1. 修复 stock_zt_pool_dtgc_em 接口，获取所有股票的数据
@@ -430,6 +434,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.5.43 fix: fix stock_zt_pool_em interface
 
 1.5.42 fix: fix stock_zt_pool_dtgc_em interface